+++ .eslintrc.json
@@ -4,7 +4,6 @@
     "react"
   ],
   "extends": [
-    "react-app",
     "eslint:recommended",
     "plugin:@typescript-eslint/recommended",
     "plugin:react/recommended"
