+++ api/global_scoreboard_api.py
@@ -1,0 +1,121 @@
+"""
+Provides the API endpoints for consuming and producing REST requests and
+responses within the Global Scoreboard context
+"""
+from api.api_wrappers import authentication_required
+from datetime import datetime, timedelta
+from flask import Blueprint, jsonify
+from models.core_models import Player
+from models.models_utils import map_to_dto
+from sqlalchemy import exc
+from typing import Dict
+from services.user_updater import get_updated_user
+from services.utils import UserUpdaterError
+import configs
+import traceback
+
+api = Blueprint('global_scoreboard_api', __name__)
+
+
+@api.route('/players', methods=('GET',))
+def get_all_players():
+    return jsonify(map_to_dto(Player.get_all()))
+
+
+@api.route('/players/<id>/score-details', methods=('GET',))
+def get_player_score_details(id: str):
+    player = Player.get(id)
+    print(player)
+    print(player.score_details)
+    if player:
+        return player.score_details or ""
+    else:
+        return "", 404
+
+
+@api.route('/players/<name_or_id>/update', methods=('POST',))
+def update_player(name_or_id: str):
+    try:
+        if configs.bypass_update_restrictions:
+            return __do_update_player_bypass_restrictions(name_or_id)
+        else:
+            return __do_update_player(name_or_id)
+    except UserUpdaterError as exception:
+        error_message = "Error: {}\n{}".format(exception.args[0]["error"], exception.args[0]["details"])
+        print(f"\n{error_message}")
+        return error_message, 424
+    except Exception:
+        error_message = "Error: Unknown\n{}".format(traceback.format_exc())
+        print(f"\n{error_message}")
+        return error_message, 500
+
+
+def __do_update_player_bypass_restrictions(name_or_id: str):
+    result = get_updated_user(name_or_id)
+    return jsonify(result), 400 if result["state"] == "warning" else 200
+
+
+__currently_updating_from: Dict[str, datetime] = {}
+__currently_updating_to: Dict[str, datetime] = {}
+
+
+@authentication_required
+def __do_update_player(current_user: Player, name_or_id: str):
+    global __currently_updating_from
+    global __currently_updating_to
+    now = datetime.now()
+
+    # Check if the current user is already updating someone
+    # or if the player to be updated is currently being updated
+    if current_user.user_id in __currently_updating_from \
+            and now - __currently_updating_from[current_user.user_id] <= timedelta(minutes=5):
+        # CONSIDER: #seconds_left = (5 * 60) - (now - __currently_updating_from[name_or_id])
+        # return {"messageKey": "current_user", "timeLeft": seconds_left}, 409
+        return "current_user", 409
+    if name_or_id in __currently_updating_to \
+            and now - __currently_updating_to[name_or_id] <= timedelta(minutes=5):
+        # CONSIDER: #seconds_left = (5 * 60) - (now - __currently_updating_to[name_or_id]).seconds
+        # return {"messageKey": "name_or_id", "timeLeft": seconds_left}, 409
+        return "name_or_id", 409
+    __currently_updating_from[current_user.user_id] = now
+    __currently_updating_to[name_or_id] = now
+
+    try:
+        # Actually do the update process
+        result = get_updated_user(name_or_id)
+    finally:
+        # Upon update completing, allow the user to update again
+        __currently_updating_from.pop(current_user.user_id, None)
+
+    return jsonify(result), 400 if result["state"] == "warning" else 200
+
+
+@api.route('/players/current/friends', methods=('GET',))
+@authentication_required
+def get_friends_current(current_user: Player):
+    return jsonify(map_to_dto(current_user.get_friends()))
+
+
+@api.route('/players/current/friends/<id>', methods=('PUT',))
+@authentication_required
+def put_friends_current(current_user: Player, id: str):
+    if current_user.user_id == id:
+        return "You can't add yourself as a friend!", 422
+    try:
+        result = current_user.befriend(id)
+    except exc.IntegrityError:
+        return f"User ID \"{id}\" is already one of your friends."
+    else:
+        if result:
+            return f"Successfully added user ID \"{id}\" as a friend."
+        else:
+            return "You can't add yourself as a friend!", 422
+
+
+@api.route('/players/current/friends/<id>', methods=('DELETE',))
+@authentication_required
+def delete_friends_current(current_user: Player, id: str):
+    if current_user.unfriend(id):
+        return f"Successfully removed user ID \"{id}\" from your friends."
+    else:
+        return f"User ID \"{id}\" isn't one of your friends."
