+++ global-scoreboard/package.json
@@ -18,7 +18,7 @@
     "@fortawesome/free-regular-svg-icons": "^5.14.0",
     "@fortawesome/free-solid-svg-icons": "^5.14.0",
     "@fortawesome/react-fontawesome": "^0.1.11",
-    "bootstrap": "^4.5.0",
+    "bootstrap": "^4.5.2",
     "react": "^16.13.1",
     "react-bootstrap": "^1.3.0",
     "react-bootstrap-table-next": "^4.0.3",
@@ -26,19 +26,25 @@
     "react-bootstrap-table2-paginator": "^2.1.2",
     "react-bootstrap-table2-toolkit": "^2.1.3",
     "react-dom": "^16.13.1",
-    "react-picky": "^5.2.5",
-    "react-scripts": "^3.4.1"
+    "react-picky": "^5.3.0",
+    "react-scripts": "^3.4.3"
   },
   "devDependencies": {
     "@types/jest": "^24.0.23",
     "@types/node": "^13.13.15",
-    "@types/react": "^16.9.43",
+    "@types/react": "^16.9.46",
     "@types/react-dom": "^16.9.8",
     "eslint": "^6.8.0",
-    "eslint-plugin-react": "^7.20.3",
+    "eslint-plugin-react": "^7.20.6",
     "tslint": "^5.20.1",
     "tslint-react": "^4.2.0",
     "typescript": "^3.9.7"
+  },
+  "eslintConfig": {
+    "extends": [
+      "../.eslintrc.json",
+      "react-app"
+    ]
   },
   "browserslist": {
     "production": [
