+++ models/core_models.py
@@ -1,20 +1,12 @@
-#!/usr/bin/python
-# -*- coding: utf-8 -*-
-
 from __future__ import annotations
-from datetime import datetime, timedelta
+from datetime import datetime
 from flask_sqlalchemy import SQLAlchemy
 from sqlalchemy import orm, text
 from typing import Any, Dict, List, Optional, Tuple, Union
-from utils import get_file, SpeedrunComError
-import configs
+from services.utils import get_file, SpeedrunComError
+import sys
 import traceback
 import uuid
-
-
-def map_to_dto(dto_mappable_object_list) -> List[Dict[str, Union[str, bool, int]]]:
-    return [dto_mappable_object.to_dto() for dto_mappable_object in dto_mappable_object_list]
-
 
 db = SQLAlchemy()
 
@@ -293,229 +285,5 @@ class Player(db.Model):
         }
 
 
-memoized_requests: Dict[str, CachedRequest] = {}
-
-
-class CachedRequest():
-    global memoized_requests
-    result: dict
-    timestamp: datetime
-
-    def __init__(self, result: dict, timestamp: datetime):
-        self.result = result
-        self.timestamp = timestamp
-
-    @staticmethod
-    def get_response_or_new(url: str) -> dict:
-        today = datetime.utcnow()
-        yesterday = today - timedelta(days=configs.last_updated_days[0])
-
-        try:
-            cached_request = memoized_requests[url]
-        except KeyError:
-            cached_request = None
-        if (cached_request and cached_request.timestamp >= yesterday):
-            return cached_request.result
-        else:
-            result = get_file(url)
-            memoized_requests[url] = CachedRequest(result, today)
-            return result
-
-
-class GameValues(db.Model):
-    __tablename__ = "game_values"
-
-    game_id: str = db.Column(db.String(8), primary_key=True)
-    category_id: str = db.Column(db.String(8), primary_key=True)
-    run_id: str = db.Column(db.String(8), nullable=False)
-    platform_id: Optional[str] = db.Column(db.String(8))
-    wr_time: int = db.Column(db.Integer, nullable=False)
-    wr_points: int = db.Column(db.Integer, nullable=False)
-    mean_time: int = db.Column(db.Integer, nullable=False)
-
-    @staticmethod
-    def create_or_update(
-            game_id: str,
-            category_id: str,
-            platform_id: Optional[str],
-            wr_time: int,
-            wr_points: int,
-            mean_time: int,
-            run_id: str):
-        existing_game_values = GameValues.get(game_id, category_id)
-        if existing_game_values is None:
-            return GameValues.create(game_id, category_id, platform_id, wr_time, wr_points, mean_time, run_id)
-        existing_game_values.platform_id = platform_id
-        existing_game_values.wr_time = wr_time
-        existing_game_values.wr_points = wr_points
-        existing_game_values.mean_time = mean_time
-        existing_game_values.run_id = run_id
-        db.session.commit()
-        return existing_game_values
-
-    @staticmethod
-    def create(
-            game_id: str,
-            category_id: str,
-            platform_id: Optional[str],
-            wr_time: int,
-            wr_points: int,
-            mean_time: int,
-            run_id: str) -> GameValues:
-        game_values = GameValues(
-            game_id=game_id,
-            category_id=category_id,
-            platform_id=platform_id,
-            wr_time=wr_time,
-            wr_points=wr_points,
-            mean_time=mean_time,
-            run_id=run_id)
-        db.session.add(game_values)
-        db.session.commit()
-
-        return game_values
-
-    @staticmethod
-    def get(game_id: str, category_id: str) -> Optional[Player]:
-        try:
-            return GameValues \
-                .query \
-                .filter(GameValues.game_id == game_id) \
-                .filter(GameValues.category_id == category_id) \
-                .one()
-        except orm.exc.NoResultFound:
-            return None
-
-    def to_dto(self) -> dict[str, Union[str, int]]:
-        return {
-            'gameId': self.game_id,
-            'categoryId': self.category_id,
-            'platformId': self.platform_id,
-            'wrTime': self.wr_time,
-            'wrPoints': self.wr_points,
-            'meanTime': self.mean_time,
-            'runId': self.run_id,
-        }
-
-
-class Schedule(db.Model):
-    __tablename__ = "schedule"
-
-    schedule_id: int = db.Column(db.Integer, primary_key=True)
-    name: str = db.Column(db.String(128), nullable=False, default='')
-    owner_id: int = db.Column(db.String(8), db.ForeignKey('player.user_id'), nullable=False)
-    registration_key: str = db.Column(db.String(36), nullable=False)
-    is_active: bool = db.Column(db.Boolean, nullable=False, default=True)
-
-    owner: Player = db.relationship("Player", back_populates="schedules")
-    time_slots: List[TimeSlot] = db.relationship(
-        "TimeSlot",
-        cascade="all,delete,delete-orphan",
-        back_populates="schedule")
-
-    @staticmethod
-    def get(id: str) -> Optional[Schedule]:
-        return Schedule.query.get(id)
-
-    @staticmethod
-    def get_with_key(schedule_id: str, registration_key: str) -> Optional[Schedule]:
-        try:
-            return Schedule \
-                .query \
-                .filter(Schedule.schedule_id == schedule_id) \
-                .filter(Schedule.registration_key == registration_key) \
-                .one()
-        except orm.exc.NoResultFound:
-            return None
-
-    def to_dto(self) -> dict[str, Union[str, int, bool, List[Dict[str, Union[str, int, bool]]]]]:
-        return {
-            'id': self.schedule_id,
-            'name': self.name,
-            'active': self.is_active,
-            'registrationKey': self.registration_key,
-            'timeSlots': map_to_dto(self.time_slots),
-        }
-
-
-class TimeSlot(db.Model):
-    __tablename__ = "time_slot"
-
-    time_slot_id: int = db.Column(db.Integer, primary_key=True)
-    schedule_id: int = db.Column(db.String(8), db.ForeignKey('schedule.schedule_id'), nullable=False)
-    date_time: datetime = db.Column(db.DateTime, nullable=False)
-    maximum_entries: int = db.Column(db.Integer, nullable=False)
-    participants_per_entry: int = db.Column(db.Integer, nullable=False)
-
-    schedule: Schedule = db.relationship("Schedule", back_populates="time_slots")
-    registrations: List[Registration] = db.relationship(
-        "Registration",
-        cascade="all,delete,delete-orphan",
-        back_populates="timeslot")
-
-    @staticmethod
-    def get_with_key(time_slot_id: str, registration_key: str) -> Optional[TimeSlot]:
-        try:
-            parent_schedule: Schedule = Schedule \
-                .query \
-                .filter(Schedule.registration_key == registration_key) \
-                .one()
-            return TimeSlot \
-                .query \
-                .filter(TimeSlot.time_slot_id == time_slot_id) \
-                .filter(TimeSlot.schedule_id == parent_schedule.schedule_id) \
-                .one()
-        except orm.exc.NoResultFound:
-            return None
-
-    def register_participant(self, participant_names: List[str]) -> int:
-        new_registration = Registration(time_slot_id=self.time_slot_id)
-        db.session.add(new_registration)
-        db.session.flush()
-
-        new_participants = [Participant(
-            registration_id=new_registration.registration_id,
-            name=participant_name)
-            for participant_name in participant_names]
-
-        db.session.bulk_save_objects(new_participants)
-
-        db.session.commit()
-        return new_registration.registration_id
-
-    def to_dto(self) -> dict[str, Union[str, int, bool, datetime, List[Dict[str, Union[str, int, bool]]]]]:
-        return {
-            'id': self.time_slot_id,
-            'dateTime': self.date_time,
-            'maximumEntries': self.maximum_entries,
-            'participantsPerEntry': self.participants_per_entry,
-            'registrations': map_to_dto(self.registrations)
-        }
-
-
-class Registration(db.Model):
-    __tablename__ = "registration"
-
-    registration_id: int = db.Column(db.Integer, primary_key=True)
-    time_slot_id: int = db.Column(db.Integer, db.ForeignKey('time_slot.time_slot_id'), nullable=False)
-
-    timeslot: TimeSlot = db.relationship("TimeSlot", back_populates="registrations")
-    participants: List[Participant] = db.relationship(
-        "Participant",
-        cascade="all,delete,delete-orphan",
-        back_populates="registration")
-
-    def to_dto(self) -> dict[str, Union[int, List[str]]]:
-        return {
-            'id': self.registration_id,
-            'participants': [participant.name for participant in self.participants]
-        }
-
-
-class Participant(db.Model):
-    __tablename__ = "participant"
-
-    registration_id: int = db.Column(db.Integer, db.ForeignKey('registration.registration_id'), primary_key=True)
-    name: str = db.Column(db.String(128), primary_key=True)
-
-    registration: Registration = db.relationship("Registration", back_populates="participants")
+if 'models.tournament_scheduler_models' not in sys.modules:
+    from models.tournament_scheduler_models import Participant, Registration, Schedule, TimeSlot
