+++ models/global_scoreboard_models.py
@@ -1,25 +1,56 @@
-#!/usr/bin/python
-# -*- coding: utf-8 -*-
+from __future__ import annotations
+from datetime import datetime, timedelta
+from typing import Dict
+from services.utils import get_file
+import configs
 
 from collections import Counter
 from datetime import datetime
 from math import ceil, exp, floor, pi
-from models import db, CachedRequest, GameValues, Player
+from models.core_models import db, Player
+from models.game_search_models import GameValues
+from models.global_scoreboard_models import CachedRequest
 from threading import Thread, active_count
 from time import strftime, sleep
 from typing import Dict, List, Optional, Tuple, Union
-from utils import UserUpdaterError, SpeedrunComError
+from services.utils import UserUpdaterError, SpeedrunComError
 import configs
 import httplib2
 import re
 import requests
 import traceback
 
-
-SEPARATOR = "-" * 64
 MIN_LEADERBOARD_SIZE = 3  # This is just to optimize as the formula gives 0 points to leaderboards size < 3
 TIME_BONUS_DIVISOR = 3600 * 12  # 12h (1/2 day) for +100%
 GAMETYPE_MULTI_GAME = "rj1dy1o8"
+
+memoized_requests: Dict[str, CachedRequest] = {}
+
+
+class CachedRequest():
+    global memoized_requests
+    result: dict
+    timestamp: datetime
+
+    def __init__(self, result: dict, timestamp: datetime):
+        self.result = result
+        self.timestamp = timestamp
+
+    @staticmethod
+    def get_response_or_new(url: str) -> dict:
+        today = datetime.utcnow()
+        yesterday = today - timedelta(days=configs.last_updated_days[0])
+
+        try:
+            cached_request = memoized_requests[url]
+        except KeyError:
+            cached_request = None
+        if (cached_request and cached_request.timestamp >= yesterday):
+            return cached_request.result
+        else:
+            result = get_file(url)
+            memoized_requests[url] = CachedRequest(result, today)
+            return result
 
 
 class Run:
@@ -253,58 +284,56 @@ class User:
             self._banned = True
             self._points = 0
 
-    def set_points(self) -> None:
+    def set_points(self, threadsException) -> None:
         counted_runs: List[Run] = []
         kept_for_game_values: List[Run] = []
 
         def set_points_thread(pb) -> None:
             try:
-                # Check if it's a valid run (has a category AND has video verification)
-                if pb["run"]["category"] and pb["run"].get("videos"):
-                    # Get a list of the game's subcategory variables
-                    url = "https://www.speedrun.com/api/v1/games/{game}/variables".format(game=pb["run"]["game"])
-                    game_variables = CachedRequest.get_response_or_new(url)
-                    game_subcategory_ids: List[str] = []
-                    for game_variable in game_variables["data"]:
-                        if game_variable["is-subcategory"]:
-                            game_subcategory_ids.append(game_variable["id"])
+                # Get a list of the game's subcategory variables
+                url = "https://www.speedrun.com/api/v1/games/{game}/variables".format(game=pb["run"]["game"])
+                game_variables = CachedRequest.get_response_or_new(url)
+                game_subcategory_ids: List[str] = []
+                for game_variable in game_variables["data"]:
+                    if game_variable["is-subcategory"]:
+                        game_subcategory_ids.append(game_variable["id"])
 
-                    pb_subcategory_variables: Dict[str, Dict[str, str]] = {}
-                    # For every variable in the run...
-                    for pb_var_id, pb_var_value in pb["run"]["values"].items():
-                        # ...find if said variable is one of the game's subcategories...
-                        if pb_var_id in game_subcategory_ids:
-                            # ... and add it to the run's subcategory variables
-                            pb_subcategory_variables[pb_var_id] = pb_var_value
+                pb_subcategory_variables: Dict[str, Dict[str, str]] = {}
+                # For every variable in the run...
+                for pb_var_id, pb_var_value in pb["run"]["values"].items():
+                    # ...find if said variable is one of the game's subcategories...
+                    if pb_var_id in game_subcategory_ids:
+                        # ... and add it to the run's subcategory variables
+                        pb_subcategory_variables[pb_var_id] = pb_var_value
 
-                    run: Run = Run(pb["run"]["id"],
-                                   pb["run"]["times"]["primary_t"],
-                                   pb["run"]["game"],
-                                   pb["run"]["category"],
-                                   pb_subcategory_variables,
-                                   pb["run"]["level"])
+                run: Run = Run(pb["run"]["id"],
+                               pb["run"]["times"]["primary_t"],
+                               pb["run"]["game"],
+                               pb["run"]["category"],
+                               pb_subcategory_variables,
+                               pb["run"]["level"])
 
-                    # Set game search data
-                    run._is_wr_time = pb["place"] == 1
-                    run._platform = pb["run"]["system"]["platform"]
+                # Set game search data
+                run._is_wr_time = pb["place"] == 1
+                run._platform = pb["run"]["system"]["platform"]
 
-                    # If a category has already been counted, only keep the one that's worth the most.
-                    # This can happen in leaderboards with coop runs or subcategories.
-                    if run._points > 0:
-                        for i, counted_run in enumerate(counted_runs):
-                            if counted_run == run:
-                                if run._points > counted_run._points:
-                                    # If other run was a WR in its sub category, keep for GameValues
-                                    if counted_run._is_wr_time:
-                                        kept_for_game_values.append(counted_run)
-                                    counted_runs[i] = run
-                                # If this run was a WR in its sub category, keep for GameValues
-                                else:
-                                    if run._is_wr_time:
-                                        kept_for_game_values.append(run)
-                                break
-                        else:
-                            counted_runs.append(run)
+                # If a category has already been counted, only keep the one that's worth the most.
+                # This can happen in leaderboards with coop runs or subcategories.
+                if run._points > 0:
+                    for i, counted_run in enumerate(counted_runs):
+                        if counted_run == run:
+                            if run._points > counted_run._points:
+                                # If other run was a WR in its sub category, keep for GameValues
+                                if counted_run._is_wr_time:
+                                    kept_for_game_values.append(counted_run)
+                                counted_runs[i] = run
+                            # If this run was a WR in its sub category, keep for GameValues
+                            else:
+                                if run._is_wr_time:
+                                    kept_for_game_values.append(run)
+                            break
+                    else:
+                        counted_runs.append(run)
 
             except UserUpdaterError as exception:
                 threadsException.append(exception.args[0])
@@ -330,8 +359,13 @@ class User:
             self._points = 0
             threads: List[Thread] = []
             for pb in pbs["data"]:
-                # Skip over "multi-game" gametype
-                if GAMETYPE_MULTI_GAME not in pb["game"]["data"]["gametypes"]:
+                # Check if it's a valid run:
+                # - not a "multi-game" gametype
+                # - has a category
+                # - has video verification
+                if GAMETYPE_MULTI_GAME not in pb["game"]["data"]["gametypes"] \
+                        and pb["run"]["category"] \
+                        and pb["run"].get("videos"):
                     threads.append(Thread(target=set_points_thread, args=(pb,)))
             for t in threads:
                 while True:
@@ -397,113 +431,3 @@ class User:
                 self._points = 0  # In case the banned flag has been set mid-thread
         else:
             self._points = 0
-
-
-def get_updated_user(p_user_id: str) -> Dict[str, Union[str, None, float, int]]:
-    """Called from flask_app and AutoUpdateUsers.run()"""
-    global threadsException
-    threadsException = []
-    text_output: str = p_user_id
-    result_state: str = "info"
-
-    try:
-        user = User(p_user_id)
-        print(f"{SEPARATOR}\n{user._name}")  # debug_str
-
-        try:
-            user.set_code_and_name()
-        except SpeedrunComError:
-            # ID doesn't exists on speedrun.com but it does in the database, remove it
-            player = Player.get(user._name)
-            if player:
-                text_output = (f"User ID \"{user._id}\" not found on speedrun.com. "
-                               "\nRemoved it from the database.")
-                result_state = "warning"
-                db.session.delete(player)
-                db.session.commit()
-            else:
-                text_output = (f"User \"{user._id}\" not found. "
-                               "\nMake sure the name or ID is typed properly. "
-                               "It's possible the user you're looking for changed their name. "
-                               "In case of doubt, use their ID.")
-                result_state = "warning"
-        else:
-            # Setup a few checks
-            player = Player.get(user._id)
-
-            # If user doesn't exists or enough time passed since last update
-            if not player or \
-                not player.last_update or \
-                (datetime.now() - player.last_update).days >= configs.last_updated_days[0] or \
-                    configs.bypass_update_restrictions:
-
-                user.set_points()
-                if not threadsException:
-
-                    print(f"\nLooking for {user._id}")  # debug_str
-                    timestamp = strftime("%Y-%m-%d %H:%M")
-
-                    # If user already exists, update the database entry
-                    if player:
-                        text_output = f"{user} found. Updated their entry."
-                        result_state = "success"
-                        Player \
-                            .query \
-                            .filter(Player.user_id == user._id) \
-                            .update({"user_id": user._id,
-                                     "name": user._name,
-                                     "country_code": user._country_code,
-                                     "score": floor(user._points),
-                                     "score_details": user._point_distribution_str,
-                                     "last_update": timestamp})
-                        db.session.commit()
-
-                    # If user not found and has points, add it to the database
-                    elif user._points >= 1:
-                        text_output = "{} not found. Added a new row.".format(user)
-                        result_state = "success"
-                        Player.create(user._id,
-                                      user._name,
-                                      country_code=user._country_code,
-                                      score=user._points,
-                                      score_details=user._point_distribution_str,
-                                      last_update=timestamp)
-                    else:
-                        text_output = f"Not inserting new data as {user} " \
-                                      f"{'is banned' if user._banned else 'has a score lower than 1.'}."
-                        result_state = "warning"
-                    text_output += user._point_distribution_str
-
-                else:
-                    error_str_list: List[str] = []
-                    for e in threadsException:
-                        error_str_list.append("Error: {}\n{}".format(e["error"], e["details"]))
-                    error_str_counter = Counter(error_str_list)
-                    errors_str = "{0}" \
-                                 "\nhttps://github.com/Avasam/Global_Speedrunning_Scoreboard/issues" \
-                                 "\nNot uploading data as some errors were caught during execution:" \
-                                 "\n{0}\n".format(SEPARATOR)
-                    for error, count in error_str_counter.items():
-                        errors_str += f"[x{count}] {error}\n"
-                    text_output += ("\n" if text_output else "") + errors_str
-                    result_state = "danger"
-            else:
-                cant_update_time = configs.last_updated_days[0]
-                text_output = f"This user has already been updated in the past {cant_update_time} day" \
-                    "s" if cant_update_time != 1 else ""
-                result_state = "warning"
-
-        return {'state': result_state,
-                'rank': None,
-                'name': user._name,
-                'countryCode': user._country_code,
-                'score': floor(user._points),
-                'lastUpdate': strftime("%Y-%m-%d %H:%M"),
-                'userId': user._id,
-                'message': text_output}
-
-    except httplib2.ServerNotFoundError as exception:
-        raise UserUpdaterError({"error": "Server not found",
-                                "details": f"{exception}\nPlease make sure you have an active internet connection"})
-    except (requests.exceptions.ChunkedEncodingError, ConnectionAbortedError) as exception:
-        raise UserUpdaterError({"error": "Connexion interrupted", "details": exception})
