+++ tournament-scheduler/package.json
@@ -15,31 +15,37 @@
   },
   "dependencies": {
     "@date-io/moment": "^1.3.13",
-    "@material-ui/core": "^4.7.1",
-    "@material-ui/icons": "^4.5.1",
-    "@material-ui/pickers": "^3.2.8",
-    "moment": "^2.24.0",
-    "react": "^16.12.0",
+    "@material-ui/core": "^4.11.0",
+    "@material-ui/icons": "^4.9.1",
+    "@material-ui/pickers": "^3.2.10",
+    "moment": "^2.27.0",
+    "react": "^16.13.1",
     "react-div-100vh": "^0.3.8",
-    "react-dom": "^16.12.0",
-    "react-scripts": "^3.4.1",
+    "react-dom": "^16.13.1",
+    "react-scripts": "^3.4.3",
     "react-text-mask": "^5.4.3"
   },
   "devDependencies": {
-    "@babel/plugin-proposal-optional-chaining": "^7.7.4",
-    "@types/jest": "^24.0.23",
-    "@types/node": "^13.11.0",
-    "@types/react": "^16.9.14",
+    "@babel/plugin-proposal-optional-chaining": "^7.11.0",
+    "@types/jest": "^24.9.1",
+    "@types/node": "^13.13.15",
+    "@types/react": "^16.9.46",
     "@types/react-div-100vh": "^0.3.0",
-    "@types/react-dom": "^16.9.4",
+    "@types/react-dom": "^16.9.8",
     "@types/react-text-mask": "^5.4.6",
-    "@typescript-eslint/eslint-plugin": "^2.14.0",
-    "@typescript-eslint/parser": "^2.14.0",
+    "@typescript-eslint/eslint-plugin": "^2.34.0",
+    "@typescript-eslint/parser": "^2.34.0",
     "eslint": "^6.8.0",
-    "eslint-plugin-react": "^7.17.0",
+    "eslint-plugin-react": "^7.20.6",
     "tslint": "^5.20.1",
     "tslint-react": "^4.2.0",
     "typescript": "3.8.2"
+  },
+  "eslintConfig": {
+    "extends": [
+      "../.eslintrc.json",
+      "react-app"
+    ]
   },
   "browserslist": {
     "production": [
